Added @id attribute contents after title in all structures like chapters, appendix, sect1, sect2, ... After switching to debug mode, the title will look like this:
1.3 A title for this section
Filename:  "foo.xml"
ID: #sec.foo

Filename and ID lines have a grey background. Filename is _always_
printed in contrast to former times. The filename is retrieved from
the @xml:base attribute. If the current node does not have such an
attribute, the XSLT stylesheet searches for a parent node which
could have that.


git-svn-id: svn+ssh://svn.code.sf.net/p/daps/svn/trunk/daps@307 2c1a2506-8097-4b87-b405-3e90b480097b