Establishing the new locdrop workflow: Step 3
Added the unpack-locdrop subcommand that takes a tarball generated with
"package-src --locdrop", an output directory and a notrans directory to
create a "daps" directory for the translated sources. Files that have been
translated are copied to the output directory, untranslated files are linked
from the notrans directory. TODO: accept more than one input tarball