Replaced FZSongTi with wqy-microhei
Needed to makes it possible to build with old suse stylesheets
without touching the suse-xsl package.
The old suse stylesheets contains a "Chinese" font name. This
font name is an alias which maps to the real fontfile in this
XEP config file.

Currently, we can only "simulate" italic, bold is not available :-(

Notes:
* FZSongTi was included in the ttf-founder-simplified package (in former
  times delivered on openSUSE 10.3)
* wqy-microhei is included in the wqy-microhei-fonts package