4 2 5 Upgrade Stylesheet: Don't put screens in CDATA
CDATA occasionally destroys things, e.g. entities in screens.
Of course that only happens because we escape entities beforehand, so
as to keep them at all.

CDATA is also not very pretty and in most cases not even helpful
(although that depends on whether there is a screen containing XML
content).

*Untested* Fingers crossed that nothing bad happens.