Fix for issue #363
Option --novalid for the target "bigfile" will create the bigfile
without performing a validation check on the sources. However, the
resulting bigfile will be validated.
Makes debugging XML errors on files with xi:includes easier. In case an
error occurs in one of the included files, the xmllint/jing error
message only points to the line in MAIN that includes the file which
contains the error. Use "bigfile --novalid" in this case. A single XML
will be built and validated. Now the error message points to the exact
location of the error in the bigfile. From there it should be easy to
find the error in the original sources.