By default, use the xep-daps.xml configuration from stylesheets
We have been shipping xep-daps.xml files for a few years with our
stylesheets now but never actually used them. This commit aims to change
that, because:

* using the XEP configuration delivered with the stylesheets, we
  can make sure we actually use the right fonts and that those
  fonts are actually installed
* we can redefine generic categories like "sans"/"serif" based on the
  stylesheets we are using and their font dependencies
* we can strip down the xep-daps.xml delivered with DAPS to a
  minimum

Note that because of a tiny XML validation issue in suse2013's
xep-daps.xml, the new default is incompatible with all
suse-xsl-stylesheets packages until 2.0.12. :(