Fix #510: Make XPointer work
Jing fails as it depends on Xerces. As Xerces does not support
the xpointer() scheme, we need to give jing a "complete"
document. The document is created by `xmllint`.

Currently it's an ugly implementation as it appends ".resolved"
to the `PROFILED_MAIN` variable. Probably we need to adapt the
code in other locations too.

For the time being, it's just the validate subcommand.