AsciiDoc multipart books can be converted into a DocBook set
An AsciiDoc dcoument consisting of multiple
parts (https://asciidoctor.org/docs/user-manual/#book-parts-and-chapter)
can be converted into a DocBook set where each part becomes a book within the
set.
Since the content is organized in a single book an the AsciiDoc-side, this
allows to link between all "books" (aka parts) within the set.

Use --adocset on the command-line or ADOC_SET="yes" in the DC-files

Requirements:
* the multipart book must only contain parts, no content outside of parts
* the book need to follow the spacial rules for Preface/Appendix headlines
  (https://asciidoctor.org/docs/user-manual/#user-preface and
  https://asciidoctor.org/docs/user-manual/#user-appendix)
* the DC-files for the individuakl books need to use the set-file as MAIN
  and the id of the respective part as ROOTID

Fixes issue#508