Asciidoc: Generate valid auto-IDs even with digits in title
By default, Asciidoctor auto-generates IDs that start with an
_ character (the idprefix). Since 305cf0988, DAPS set an empty idprefix
parameter for Asciidoctor. However, following that, whenever Asciidoctor
had to generate IDs for titles that start with, e.g. a digit, such as all
the GFDL bridgeheads, "0. PREAMBLE", ... it would generate invalid XML IDs
like "0-preamble". (Digits are not allowed at the very beginning of an
ID, neither are dashes.)

To circumvent that issue, this commit adds the prefix "id-".