Make AsciiDoc default attributes overridable
In doc-caasp, we have the situation that there are a bunch of maintenance
branches that we want to continue to keep publishable with the old ID
generation scheme (because changing all the URLs on doc.suse.com sucks).

Now, AsciiDoctor always uses the --attribute value defined last.
This made the former default definitions non-overridable. Thus not
ideal when trying to keep stable URLs across DAPS updates.

This commit splits the attributes up into an overridable set (for the IDs)
and a non-overridable (for the image directory etc.).