daps-migrate: `sentence-case` converts title case to sentence case
Sentence case looks for `<title/>` and `<term/>` tags and tries to
convert them from title case to sentence case. It uses a list of common
trademarks and proper nouns in the hope of limiting damage to such
words. While it works decently well, the process is inherently lossy
and will require some manual rework in most cases.
It tries to ignore nested tags and entities. However, it is not based
on an established XML tool.