xmlformat: Switch config file at runtime depending on xmlformat version
This is necessary as whenever xmlformat encounters an unknown
attribute/value pair, it will error out.

The versionless xmlformat config file is now a symlink to the variant
for xmlformat 1.04. This provides backward compatibility with the
behavior of DAPS 3.1.2 and below. If necessary, we can change that
symlink in the future.