python: Symlink scripts in a more practical way for packaging
Before the entire repo content for these Python scripts was integrated
into the DAPS repo, the bare scripts were in `libexec/` already. After
the move we symlinked from `libexec`->`python-scripts`.
However, we never changed any of the packaging and keeping the actual
scripts in `libexec` is in fact more practical, so let's reverse the
symlinks.