Changing detection of not optimized PNG (fixes #513)
When optimizinmg PNGs with optipng an exif comment "optipng" was added
to the image. However, when editing the image afterwards and saving it
without optimization, it would be skipped by subsequent optipng runs,
because the exif information was still there.
This patch changes the following:
* exif information is no longer needed (also removes dependency on
  exiftool)
* with each optipng run, all images are scanned by optipng (although
  only non-optimzed images are changed). This comes with a significant
  performance penalty
* "optipng -o0 -simulate" is used to check whther an image is already
  optimized (no performance penalty compared to the exiftoll method)

Also makees sure images that cause iCCP warnings with libpng > 1.6 get
automatically fixed (optipng -fix).