Added new option for validation: --extended-validation
--extended-validation=value
    Comma-separated list of extended validation checks
     Possible values:
       all|ALL: shortcut for all checks listed below
       ids:     checks for IDs that are not SEO friendly
       images:  checks for missing and duplicated  images
       inlines: checks for empty inline elements
       tables:  checks for invalid tables
     Example: --extended-validation="ids,tables"
     Default: unset

In a DC-file these options can be activated by setting the variable
EXTENDED_VALIDATION, for example

EXTENDED_VALIDATION="ids,tables"
(the "ALL" shortcut works here as well)

Whenever one of the activated extended validation options fails, DAPS exits
with error.

This option is available for the following subcommands

* epub
* html
* locdrop
* man
* package-html
* package-pdf
* package-src
* pdf
* text
* validate
* webhelp

NOTE:
Previously, once a successful validation was done, the code was only
re-validated when the sources had changed. With the introduction of this
option the code is always re-validated (for the subcommands mentioned above).