Fixed pdf static dir handling (issue #719)
Simplified and fixed handling of admonition and callout graphics in PDFs.
The following locations for these graphics is supported:

STYLEIMG=

 1. <STYLEROOT>/fo/static/images
 2. <STYLEROOT>/fo/images
 3. <STYLEROOT>/static/images (suse-xsl-stylesheets)
 4. <STYLEROOT>/images (DocBook upstream)

First one (in the order above) wins if multiple of these locations exist.
It is assumed that the image directory layout follows DocBook upstream where
the admonition images reside directly under $STYLEIMG and the callouts under
$STYLEIMG/callouts/.

Note that in DocBook upstream admonitions are disabled by default and that low
resolution PNGs are used when acticvated. Consider using the following
parameters on the DAPS command line or use PARAM/STRINGPARAM in the DC-file to
override:

-param="admon.graphics=1"
--stringparam="admon.graphics.extension=.svg"