Stylesheet to add missing <meta> into <info>
The stylesheet looks if a specific meta tag is available. If it's available,
it copies the tag to the result tree unchanged.
If the respective tag is not available, the stylesheet creates the tag in the
result tree. In some cases you can pass a parameter to add the content (text or
child elements) to the tag.