Remove all xml:base attributes from the XML input (#741)
When using xmlstarlet ed -d '//@xml:base', the result XML is formatted.
Unfortunately, this destroys the indentation for <screen> elements.
As xmlstarlet doesn't have any options to avoid that, we can only use
a XSLT stylesheet.

This stylesheet is an attempt to replace above "xmlstarlet ed" but
preserves the indendation.